refactor : 사이드바 아이콘 svg 분리 및 디자인 통일
